refactor: make GeneralTransportExplorer
